Add files via upload
This is a dataset from the Mongolian Statistical Microdata. The Reproductive Health Survey was conducted by NSO for 3 times, the previous two times were in 1998 and 2003. As it is stated on the website "The importance of this survey is to provide data for a wide range of activities, including monitoring the indicators of the Millenium Development Goals, evaluating policies at the international and national levels, assessing required measures related to programme implementation in the future, and conducting additional in-depth research and analysis". The survey's purpose was to gather data on the reproductive health indicators such as fertility, infant and child mortality, factors affecting mortality, child health, family planning and STI/HIV/AIDS, women's knowledge on reproductive diseases, etc.

I wanted to know the situation of women in Mongolia, especially those who have children. There were three datasets in this survey, but I decided to use only one dataset women. I don't think I am able to fully use all the data provided in this survey, but hopefully in the future I will be able to do that.

I chose the Reproductive Health Survey 2008 from the Microdata of National Statistics Office of Mongolia